requirements.txt
strands-agents
strands-agents-tools
requests
fastapi
uvicorn[standard]
python-multipart
jinja2
python-dotenv